+++ pkg/cmd/task/create_test.go
@@ -1,16 +1,10 @@
 package task
 
 import (
-	"context"
-	"crypto/tls"
 	"io"
-	"net"
-	"net/http"
-	"net/http/httptest"
 	"testing"
 	"time"
 
-	"github.com/google/go-cmp/cmp"
 	"github.com/jonboulle/clockwork"
 	"github.com/tektoncd/cli/pkg/test"
 	cb "github.com/tektoncd/cli/pkg/test/builder"
@@ -64,7 +58,7 @@ func TestTaskCreate(t *testing.T) {
 			input:       seeds[1],
 			inputStream: nil,
 			wantError:   true,
-			want:        "does not support such extension for ./testdata/task.txt",
+			want:        "the path ./testdata/task.txt does not match target pattern",
 		},
 		{
 			name:        "Mismatched resource file",
@@ -108,76 +102,4 @@ func TestTaskCreate(t *testing.T) {
 			}
 		})
 	}
-}
-
-func TestLoadRemoteFile(t *testing.T) {
-	h := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		okResponse := `
-apiVersion: tekton.dev/v1alpha1
-kind: Task
-metadata:
-  name: build-docker-image-from-git-source
-spec:
-  inputs:
-    resources:
-      - name: docker-source
-        type: git
-    params:
-      - name: pathToDockerFile
-        type: string
-        description: The path to the dockerfile to build
-        default: /workspace/docker-source/Dockerfile
-      - name: pathToContext
-        type: string
-        description:
-          The build context used by Kaniko
-          (https://github.com/GoogleContainerTools/kaniko#kaniko-build-contexts)
-        default: /workspace/docker-source
-  outputs:
-    resources:
-      - name: builtImage
-        type: image
-  steps:
-    - name: build-and-push
-      image: gcr.io/kaniko-project/executor:v0.9.0
-      # specifying DOCKER_CONFIG is required to allow kaniko to detect docker credential
-      env:
-        - name: "DOCKER_CONFIG"
-          value: "/builder/home/.docker/"
-      command:
-        - /kaniko/executor
-      args:
-        - --dockerfile=$(inputs.params.pathToDockerFile)
-        - --destination=$(outputs.resources.builtImage.url)
-        - --context=$(inputs.params.pathToContext)`
-		_, _ = w.Write([]byte(okResponse))
-	})
-	httpClient, teardown := testingHTTPClient(h)
-	defer teardown()
-
-	cli := NewClient(time.Duration(0))
-	cli.httpClient = httpClient
-	remoteContent, _ := loadFileContent("https://foo.com/task.yaml", cli)
-
-	localContent, _ := loadFileContent("./testdata/task.yaml", nil)
-
-	if d := cmp.Diff(remoteContent, localContent); d != "" {
-		t.Errorf("Unexpected output mismatch: %s", d)
-	}
-}
-
-func testingHTTPClient(handler http.Handler) (*http.Client, func()) {
-	s := httptest.NewTLSServer(handler)
-
-	cli := &http.Client{
-		Transport: &http.Transport{
-			DialContext: func(_ context.Context, network, _ string) (net.Conn, error) {
-				return net.Dial(network, s.Listener.Addr().String())
-			},
-			TLSClientConfig: &tls.Config{
-				InsecureSkipVerify: true,
-			},
-		},
-	}
-	return cli, s.Close
 }
